update buy now

+++ checkout_query.php
@@ -9,6 +9,8 @@
         $phone = $_POST['phone'];
         $alamat = $_POST['alamat'];
         $pemesanan_makanan = $_POST['pemesanan_makanan'];
+        $qty_buynow = $_POST['qty_buy_now'];
+        $id_buynow = $_POST['id_buynow'];
 
         // Mendapatkan data user untuk mengupdate alamat user jika alamat tersebut masih bernilai kosong/NULL
         $getUserData = "SELECT * FROM user WHERE phone = '{$_SESSION['data']['phone']}';";
@@ -23,19 +25,25 @@
             // Mengambil ID yang baru saja dibuat pada tabel pembelian
             $lastInsertedId = mysqli_insert_id($conn);
 
-            if($lastInsertedId) { 
-                // Mengambil barang dari cart
-                $getProductFromCart = "SELECT * FROM cart WHERE id_user = '{$_SESSION['data']['phone']}'";
-                $getProductFromCartQuery = mysqli_query($conn, $getProductFromCart);
+            if($lastInsertedId) {
+                if($id_buynow !== "" || NULL) {
 
-                while($data = mysqli_fetch_array($getProductFromCartQuery)) {
-                    // Memindah barang dari cart ke dalam barang_dibeli dan menghapus tersebut dari cart
-                    $barangDibeli = "INSERT INTO barang_dibeli(id_pembelian, id_barang, jumlah) VALUES('$lastInsertedId', '{$data['id_barang']}', '{$data['jumlah']}');";
+                    $barangDibeli = "INSERT INTO barang_dibeli(id_pembelian, id_barang, jumlah) VALUES('$lastInsertedId', '$id_buynow', '$qty_buynow');";
                     $barangDibeliQuery = mysqli_query($conn, $barangDibeli);
+                } else {
+                    // Mengambil barang dari cart
+                    $getProductFromCart = "SELECT * FROM cart WHERE id_user = '{$_SESSION['data']['phone']}'";
+                    $getProductFromCartQuery = mysqli_query($conn, $getProductFromCart);
 
-                    // Check jika barang telah berhasil ditambahkan ke dalam tabel barang_dibeli maka product dalam cart akan dihapus
-                    if($barangDibeliQuery) {
-                        mysqli_query($conn, "DELETE FROM cart WHERE id = '{$data['id']}';");
+                    while($data = mysqli_fetch_array($getProductFromCartQuery)) {
+                        // Memindah barang dari cart ke dalam barang_dibeli dan menghapus tersebut dari cart
+                        $barangDibeli = "INSERT INTO barang_dibeli(id_pembelian, id_barang, jumlah) VALUES('$lastInsertedId', '{$data['id_barang']}', '{$data['jumlah']}');";
+                        $barangDibeliQuery = mysqli_query($conn, $barangDibeli);
+
+                        // Check jika barang telah berhasil ditambahkan ke dalam tabel barang_dibeli maka product dalam cart akan dihapus
+                        if($barangDibeliQuery) {
+                            mysqli_query($conn, "DELETE FROM cart WHERE id = '{$data['id']}';");
+                        }
                     }
                 }
             }
